test+log(retrieval): move validation tests to unit suite + log mental_model filter no-op (round-8 review)
- Move test_invalid_intent_class_rejected and test_invalid_risk_class_rejected
  out of integration class to a new unit test file. They were silently skipped
  by pytest -m "not integration" CI runs because the parent class carried the
  integration marker, even though the tests only instantiate RetrievalRequest
  with no DB or async I/O.
- Add a logger.debug call in MentalModelStrategy.get_statement when
  intent_class/risk_class kwargs are present, since the filter is intentionally
  not applied for mental_model retrieval (out-of-scope per issue #92). Without
  this trace, callers got a silent no-op invisible in production logs.
