+++ packages/core/tests/integration/memory/retrieval/test_int_intent_risk_filter.py
@@ -277,14 +277,6 @@ class TestIntentRiskFilter:
         ):
             assert seeded_units[key] in result_ids
 
-    async def test_invalid_intent_class_rejected(self):
-        with pytest.raises(ValueError, match='Invalid intent_class'):
-            RetrievalRequest(query='x', intent_class='not-a-real-class')
-
-    async def test_invalid_risk_class_rejected(self):
-        with pytest.raises(ValueError, match='Invalid risk_class'):
-            RetrievalRequest(query='x', risk_class='not-a-real-class')
-
 
 @pytest.mark.integration
 class TestIntentRiskNullSemantics:
